v2.1.2: Rename some files
